[core] Deprecate LIFO/FIFO worker killing policies (#56314)
Deprecating lifo/fifo worker killing policies as its unlikely for a user
to use anything other than group by owner.

Also fixing an issue where the attempt number/max retries and actor
restart number/max restarts are different types in the proto. Also
fixing an issue where the numbering of the ActorCreationTask proto
started at 2 rather than 1.

---------

+++ src/ray/common/task/task_spec.cc
@@ -187,7 +187,7 @@ bool TaskSpecification::HasRuntimeEnv() const {
   return !IsRuntimeEnvEmpty(SerializedRuntimeEnv());
 }
 
-uint64_t TaskSpecification::AttemptNumber() const { return message_->attempt_number(); }
+int32_t TaskSpecification::AttemptNumber() const { return message_->attempt_number(); }
 
 bool TaskSpecification::IsRetry() const { return AttemptNumber() > 0; }
